backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
tensorflow==2.15.0
Pillow==10.1.0
pytesseract==0.3.10
numpy==1.24.3
pymongo==4.6.1
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
bcrypt==4.1.2
pytest==7.4.3
openai==1.3.0
